Add files via upload
"showImageBitPlanes" will dissect an 8-bit image to show each of its bitplanes. (MSB will be on index 1 and so on)
"MSEandPSNR" will take two images and show the Mean Squared Error and Peak Signal-to-Noise Ration.
"fourMSBtoImage" this will take bitplanes (generated by "showImageBitPlanes") and reconstructs the image with only four most significant bits.